services/api-gateway/requirements.txt
quart==0.19.4
hypercorn==0.15.0
sqlalchemy==2.0.27
alembic==1.13.1
pydantic==2.6.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
redis==5.0.1
celery==5.3.6
python-dotenv==1.0.1 